Move util modules to sub-package
