+++ Source/cielim-python/cielim/utils/image_comparison_toolkit.py
